2001-02-16  Greg Bumgardner <bumgard@roguewave.com>
	libstdc++/1734
        * include/bits/std_ostream.h: Replaced usage of
        _Traits::_S_eos() with _CharT() as per section 17.2.2.1.3.
        * include/bits/char_traits.h: Removed non-standard
        methods that cannot be used elsewhere.

	libstdc++/1885
        * include/bits/basic_ios.h: Uncommented #include of
        bits/basic_ios.tcc

	libstdc++/1897
        * include/bits/codecvt.h: See next...
        * include/bits/fstream.tcc: See next...
        * include/bits/ostream.tcc: Add __extension__
        to variable-length arrays.

	libstdc++/1967
        * include/bits/localefwd.h: Changed int counter type to size_t.

	libstdc++/1968
        * include/bits/std_fstream.h: Added typedef for __ctype_type.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@39802 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/include/bits/localefwd.h
@@ -397,7 +397,7 @@ namespace std
     {
       _M_impl = new _Impl(*__other._M_impl, 1);
       _M_impl->_M_install_facet(&_Facet::id, __f);
-      for (int __i = 0; __i < _S_num_categories; ++__i)
+      for (size_t __i = 0; __i < _S_num_categories; ++__i)
 	_M_impl->_M_names[__i] = "*";
     }
